requirements.txt
# ============================================
# AI Resume Screener - Dependencies
# ============================================

# --- Web Scraping ---
selenium>=4.15.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
requests>=2.31.0

# --- PDF/Document Parsing ---
pdfminer.six>=20221105
PyPDF2>=3.0.0
python-docx>=1.1.0

# --- NLP ---
spacy>=3.7.0
nltk>=3.8.0

# --- ML/Vectorization ---
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0

# --- API ---
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6

# --- Data Validation ---
pydantic>=2.5.0
pydantic-settings>=2.1.0

# --- Testing ---
pytest>=7.4.0
pytest-asyncio>=0.21.0

# --- Utils ---
python-dotenv>=1.0.0
tqdm>=4.66.0
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.23.0
